server/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pandas
openpyxl
faiss-cpu
sentence-transformers
langchain==0.2.1
langchain-core==0.2.1
langchain-community==0.2.1
huggingface-hub
transformers
python-multipart
